01.chatbot_langchain_openai/app/requirements.txt
langchain
langchain-openai
python-dotenv
fastapi
uvicorn